Migrated to spring framework

+++ src/main/java/me/nasukhov/TukitaLearner/study/StudentStats.java
@@ -1,4 +1,4 @@
 package me.nasukhov.TukitaLearner.study;
 
-public record StudentStats(String studentId, int correctAnswers, int incorrectAnswers) {
+public record StudentStats(String studentId, long correctAnswers, long incorrectAnswers) {
 }
